+++ config/config.php
@@ -18,30 +18,40 @@ return [
     */
 
     'auth' => [
-        'driver' => 'database',
+        'driver' => 'database'
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | Bot configuration
+    |--------------------------------------------------------------------------
+    */
+
+    'bot' => [
         'languages' => ['en']
     ],
 
     /*
     |--------------------------------------------------------------------------
-    | Main bot configuration
+    | Main configuration
     |--------------------------------------------------------------------------
     */
 
-    'bot' => [
+    'env' => [
         'token' => env('LARAGRAM_BOT_TOKEN'),
-        'prefix' => env('LARAGRAM_WEBHOOK_PREFIX', 'bot'),
+        'prefix' => env('LARAGRAM_WEBHOOK_PREFIX', 'laragram'),
         'secret' => env('LARAGRAM_WEBHOOK_SECRET', null),
     ],
 
     /*
     |--------------------------------------------------------------------------
-    | View bot configuration
+    | Paths configuration
     |--------------------------------------------------------------------------
     */
 
-    'view' => [
-        'path' => 'bot'
-    ]
+    'paths' => [
+        'route' => 'laragram',
+        'views' => 'laragram'
+    ],
 
 ];