+++ src/Console/LaragramPublishCommand.php
@@ -53,6 +53,8 @@ class LaragramPublishCommand extends Command
 
             $this->publishViews();
 
+            $this->publishLangs();
+
             $this->publishControllers();
 
             $this->publishRoutes();
@@ -68,9 +70,29 @@ class LaragramPublishCommand extends Command
      */
     protected function createDirectories()
     {
-        if (!is_dir($directory = resource_path(config('laragram.view.path')))) {
+        if (!is_dir($directory = resource_path(config('laragram.paths.views')))) {
             mkdir($directory, 0755, true);
         }
+
+        if (!is_dir($directory = app_path('Http/Controllers/Laragram'))) {
+            mkdir($directory, 0755, true);
+        }
+    }
+
+    /**
+     * Publish the langs.
+     *
+     * @return void
+     */
+    public function publishLangs()
+    {
+        if (file_exists($file = app('path.lang') . '/en/laragram.php') && !$this->option('force')) {
+            if (!$this->confirm("The [{$file}] controller already exists. Do you want to replace it?")) {
+                return;
+            }
+        }
+
+        copy(__DIR__ . '/stubs/lang/en/laragram.stub', $file);
     }
 
     /**
@@ -81,7 +103,7 @@ class LaragramPublishCommand extends Command
     protected function publishViews()
     {
         foreach ($this->views as $key => $value) {
-            if (file_exists($view = resource_path(config('laragram.view.path') . '/' . $value)) && !$this->option('force')) {
+            if (file_exists($view = resource_path(config('laragram.paths.views') . '/' . $value)) && !$this->option('force')) {
                 if (!$this->confirm("The [{$value}] view already exists. Do you want to replace it?")) {
                     continue;
                 }
@@ -98,7 +120,7 @@ class LaragramPublishCommand extends Command
      */
     protected function publishControllers()
     {
-        if (file_exists($file = app()->path() . '/Http/Controllers/BotController.php') && !$this->option('force')) {
+        if (file_exists($file = app()->path() . '/Http/Controllers/Laragram/HelloController.php') && !$this->option('force')) {
             if (!$this->confirm("The [{$file}] controller already exists. Do you want to replace it?")) {
                 return;
             }
@@ -108,7 +130,7 @@ class LaragramPublishCommand extends Command
     }
 
     /**
-     * Compiles the BotController stub.
+     * Compiles the HelloController stub.
      *
      * @return string
      */
@@ -117,7 +139,7 @@ class LaragramPublishCommand extends Command
         return str_replace(
             '{{namespace}}',
             $this->getAppNamespace(),
-            file_get_contents(__DIR__ . '/stubs/controllers/BotController.stub')
+            file_get_contents(__DIR__ . '/stubs/controllers/HelloController.stub')
         );
     }
 
@@ -138,7 +160,7 @@ class LaragramPublishCommand extends Command
      */
     protected function publishRoutes()
     {
-        if (file_exists($file = base_path('routes/laragram.php'))) {
+        if (file_exists($file = base_path('routes/' . config('laragram.paths.route') . '.php'))) {
             file_put_contents($file, file_get_contents(__DIR__ . '/stubs/routes/laragram.stub'), FILE_APPEND);
         }
     }